Added File Uploading.
This version gives the user more detailed information about the virus itself, instead of only say if the URL itself is malicous and that's it.
A new feature was added, in this case is a early stage of file uploading. This feature will work much better once i add the entire tool a GUI for a easier use for non-experienced users.
In the future, those URL's and files will get check by more antivirus and will be filtered with more options for a safer usage.